net: dsa: microchip: add GPIOs to chip infos
The number of periodic outputs available is hardcoded to 3. KSZ8463 has
up to 12 GPIO that can be used as periodic outputs.

Add num_gpio attribute to the struct ksz_chip_data that tells how many
GPIOs are available.
Set the new attribute to 12 for the KSZ8463.
Set the new attribute to 3 for the other PTP capable switches.

Signed-off-by: Bastien Curutchet (Schneider Electric) <tux@bootlin.com>